+++ src/main/java/nl/surfnet/bod/idd/IddOfflineClient.java
@@ -58,7 +58,7 @@ public class IddOfflineClient implements IddClient {
    * {@inheritDoc}
    */
   @Override
-  public Klanten getKlantById(final int klantId) {
+  public Klanten getKlantById(final Long klantId) {
     Klanten matchedKlant = null;
 
     Collection<Klanten> klanten = getKlanten();
